:truck: Restructured mark data and request
