--- base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/data/climb/path/MarkingDataInt.kt
@@ -1,3 +1,0 @@
-package com.arnyminerz.escalaralcoiaicomtat.data.climb.path
-
-interface MarkingDataInt
